Merged PR 3121529: Fix remaining issues with Conpty Optimizations
This restores the previously reverted conpty optimizations, and fixes the few issues with them.

The diff from the previous PR is in three places:
`getset.cpp:425` - Don't invalidate all on a output mode change in conpty mode
`src/renderer/vt/paint.cpp:450, 480, 513`: use the actual number of columns for a character, not the num characters, and when we write out spaces at the end of a string, advance the logical cursor.

Cherry picked from !3120119

Related work items: #20297559, #21091917