Merged PR 3130098: Merged PR 3113196: Implement custom layout/rendering pipeline for DirectX
Merged PR 3113196: Implement custom layout/rendering pipeline for DirectX

This uses several SDK examples and consultation with the DirectWrite and Direct2D team to build us a custom layout and rendering engine.

We're doing this to ensure that we can manipulate individual characters, glyphs, and runs however we want. We can also perform additional steps during analysis like realizing that font fallback is occuring, substitute certain glyph points with others, scale things by width and height, and do it all in a route that offers us maximum performance (because it ignores all the concerns that the generalized renderer and layout have that do not apply in a monospaced cell-based rendering context like ours.)

I make no guarantees that this is perfect, but it is an implementation that accurately replaces the defaults and works like the previous one for fallback to color emoji, fallback to Japanese in a latin string, and works with the Fira Code ligatures.

Related work items: #20956925

Related work items: #20956925