Working Directory Support for Profiles
This PR introduces support for Terminal users to declare a working directory for profiles in their profiles.json. The working directory is considered optional, so users do not need to explicitly supply one. If one is supplied with a given profile though, this code will expand any environment variables present in the string, do some basic validation, and use it as the working directory for that specific session.

How verified: manually tested various combinations of valid / invalid inputs, and verified correctness of the resulting behavior.

Related work items: #20641978