Merged PR 3131441: Make Cascadia work with Per-MonitorV2 high DPI
This turns PerMonitorV2 high DPI back on for Cascadia.

To get this to work, I've moved the area of concern for DPI down from the Terminal Control into the DX Renderer. This matches the separation of control from other popular OS components that get this right including Cortana, video playback controls, Bing Maps, and even DWM itself.

The Terminal Control has been relegated to just shunting information about the panel pixel size and the scaling factor into the renderer. It has been relieved of storing any information about DPI persistently and ignores/skips DPI math in most of its calculations now as it will be handled at a different level.

The Renderer code is updated to understand that for Composition scenarios, we need to scale up the drawing surface available so more pixels are on the backend, set the scaling factor on the drawing target so everything is drawn to scale with the same device-independent-pixel calls as the regular DPI route, not blow up the font artificially (like we do for hwnd/conhost surfaces), and inform the swap chain of the inverse matrix of the scaling for final presentation onto the appropriate size panel area.

Also included is a little optimization to the Renderer code that will simply resize the swap chain and recreate the rendering target instead of rebuilding the entire device-dependent object pipeline when the display surface pixel count changes.

Related work items: #20641961, #20642286