Merged PR 3148155: Fix the fuzzy-ness seen when using Fira Code at small resolutions
This didn't afflict just Fira Code, it was a widespread issue that is exacerbated by Fira Code specifically and small font sizes generally.

I talked with the DirectWrite team and the answer is to ensure that our "baseline" for the text is always an integer value such that we're not trying to draw vertically on fractional pixels.

Since we always want an integer pixel size box in height, I just rounded up both the ascent and descent portions of the font metrics to get a full integer size and placed the baseline at the integer ascent height. There's more details on what all that means in the comments embedded in the code.

I removed two variables that were storing state in the renderer that were never used again related to font information.

I also took this opportunity to make the DirectX renderer treat the requested font size value as points instead of pixels. Generally every application you use treats "12" as points and scales appropriately, but we've always done pixels. In the interest of finally being like everyone else and ensuring our text looks like everyone else's (for the most part), I've applied the scaling factor to turn pixels into points.

Finally, to preserve the functionality of the property sheet in conhost.exe and in other codepaths that are copying the existing font into a new Desired/proposed font, I'm now just passing back the original requested "font size" as the unscaled size because this renderer does so much more work and adjustment to "scale" a requested size into something actually workable for our environment.

The scaled size will remain the pixel perfect cell grid count in each dimension, which I believe is only used by things adapting pixel positions to cell buffer grid positions.

Related work items: #21183156