Merged PR 3165394: Merged PR 3129098: Use the Initial Size settings to create the window
Merged PR 3129098: Use the Initial Size settings to create the window

This changes enables Cascadia to use the values in the default profile's `initialRows` and `initialCols` settings to create the window at the right size to create a terminal with that size. It will create a temporary DX renderer to ask it how big the font of the default profile is, and use that to get the right window size.

Related work items: #20817473

Related work items: #20817473