Merged PR 3166603: Merged PR 3165404: Switch to a Single Threaded Apartment
Merged PR 3165404: Switch to a Single Threaded Apartment

This introduces some magic load-bearing code, to switch us to a single-threaded apartment. We need the main thread to be a STA for the clipboard APIs to work correctly, but the Storage APIs specifically require that their thread is a MTA. Unfortunately, we need to call the Storage APIs synchronously on the main thread, _before_ we set up an event dispatcher.

We can get around this problem by calling the Storage API's *before* we call `winrt::init_apartment(single)`. We can't call `init_apartment` more than once, but until we call it, winrt will assume we're a MTA. So, we load the settings, then we'll initialize ourself to a STA.

We know this works because Kenny Kerr said it would, and Kenny is our god.

This also has a fix for us crashing when we launch the settings json.

Related work items: #21226546

Related work items: #21226546