Merged PR 3178992: Enable setting for scrollbarState
This PR adjusts the existing setting for the scrollbar state to be an enum, and applies it to the control based on the user's desired behavior. The enum only ended up having two states based on how the platform behaves (in order to have the scrollbar always shown, we rely on the OS setting which controls visibility of scrollbars).

Related work items: #21084789