Merged PR 3190529: NuGet package maintenence & removing duplicate TAEFs
We had two TAEFs.
1. In the dep directory
1. Coming in from the internal TAEF NuGet feed (Taef.Redist.wlk)

Now we have just the NuGet package one. I've updated all references in the entire project/solution to point to the nuget location, updated the test common to reference it (updated the C# UIA test separately because it doesn't use common), and updated all the scripts I could find in the tools directory to reference it.

In addition, it's kinda difficult for people on the outside to pull our internal-only NuGet feeds. So I've instead dropped the two .nupkgs into the ./dep/packages folder and updated the NuGet.config to look there until we get those partners to actually provide public nuget feed sources.

This should be OK in the interim because (as far as I'm aware) it's OK to redist the Taef.Redist.Wlk and there's theoretically a plan one day to get the private dev branch tabs stuff from Windows.UI.XAML merged back into the public feed. (I left a `README.md` to this effect in the new directory.)

Finally: BONUS FIX
There are too many project files in the solution so I moved them all into folders to help us mentally organize it better when viewing inside Visual Studio.

Related work items: #21343890